The hunt for the dashed line (negative weight) bug continues...
git-svn-id: https://svn.code.sf.net/p/socnetv/code@211 fc2d5e9b-b3ce-4657-a66d-b27cdf764933

+++ trunk/src/edge.h
@@ -68,7 +68,7 @@ public:
 	QString color() ;
 	void addWeight (EdgeWeight* canvasWeight  ) ;
 	void clearWeightList();
-	int lineWidth(); 
+	qreal lineWidth(); 
 	void showArrows(bool);
 	void toggleAntialiasing(bool);
